Move a use case from Model to Activity
